farmai-backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
python-dotenv==1.0.0
numpy==1.24.3
pandas==2.0.3
pillow==10.0.0
opencv-python==4.7.0.72
earthengine-api==0.1.350
scikit-learn==1.2.2
geemap==0.20.2
pydantic==1.10.7